refactor: google api를 사용하지 않기 때문에 필요한 클라이언트 측 변경 사항 적용
