+++ client/src/components/RestaurantDetail/RestaurantDetailCarousel/styles.tsx
