+++ client/src/components/RestaurantRow/index.tsx
@@ -23,7 +23,7 @@ interface PropsType {
 }
 
 function RestaurantRow({ restaurant, restaurantListType }: PropsType) {
-  const { name, category, lat, lng, rating, photoKeyList } = restaurant;
+  const { name, category, lat, lng, rating, photoUrlList } = restaurant;
 
   const {
     meetLocation: { lat: roomLat, lng: roomLng },
@@ -32,16 +32,12 @@ function RestaurantRow({ restaurant, restaurantListType }: PropsType) {
   const straightDistance = getDistance({ lat, lng }, { lat: roomLat, lng: roomLng });
 
   // 커밋에 기록이 남는게 찝찝해서 하드코딩하여 테스트했던 api key 일단 삭제.
-  const googleApiKey = '';
-
-  const photoReference = photoKeyList && photoKeyList.length > 0 ? photoKeyList[0] : '';
-
-  const imageSrc = `https://maps.googleapis.com/maps/api/place/photo?maxwidth=400&photo_reference=${photoReference}&key=${googleApiKey}`;
+  const thumbnailSrc = photoUrlList && photoUrlList[0] ? photoUrlList[0] : '';
 
   return (
     <RestaurantRowBox>
       <ImageBox>
-        <ThumbnailImage src={imageSrc} />
+        <ThumbnailImage src={thumbnailSrc} />
       </ImageBox>
       <InfoBox>
         <NameBox>{name}</NameBox>
